refactor(agent): regroup runner into shared/ + symmetric linear/orchestrator modes
Moves agent/orchestrator/ into runner/orchestrator/ and runner/{bootstrap,types,errors}.ts
into runner/shared/, giving the 4-concern layout: index (dispatch), shared/, linear, orchestrator/.
Fixes all broken imports; shared/ uses @lib/@utils/@ui aliases to stay flat.
