+++ src/lib/agent/runner/orchestrator/__tests__/executor.test.ts
@@ -5,8 +5,11 @@ import {
   QueueStore,
   type QueuedTask,
   type TaskHandoff,
-} from '@lib/agent/orchestrator/queue';
-import { drainQueue, type RunTask } from '@lib/agent/orchestrator/executor';
+} from '@lib/agent/runner/orchestrator/queue';
+import {
+  drainQueue,
+  type RunTask,
+} from '@lib/agent/runner/orchestrator/executor';
 
 jest.mock('@utils/analytics', () => ({
   analytics: { captureException: jest.fn(), wizardCapture: jest.fn() },
