+++ src/lib/agent/runner/orchestrator/__tests__/queue.test.ts
@@ -6,7 +6,7 @@ import {
   QUEUE_DIR_NAME,
   type QueueFile,
   type TaskHandoff,
-} from '@lib/agent/orchestrator/queue';
+} from '@lib/agent/runner/orchestrator/queue';
 
 jest.mock('@utils/analytics', () => ({
   analytics: { captureException: jest.fn(), wizardCapture: jest.fn() },
