+++ src/lib/agent/runner/orchestrator/__tests__/run-metrics.test.ts
@@ -1,4 +1,4 @@
-import { RunMetrics } from '@lib/agent/orchestrator/run-metrics';
+import { RunMetrics } from '@lib/agent/runner/orchestrator/run-metrics';
 
 describe('RunMetrics', () => {
   it('reports time to first start and first completion from run start', () => {
